+++ src/hooks/index.ts
@@ -1,0 +1,3 @@
+import usePagination from './pagination.hook';
+
+export { usePagination };
