+++ src/views/service/MyPage/MyPageRentList/UserRentList/index.tsx
@@ -1,13 +1,14 @@
-import React, { ChangeEvent, KeyboardEvent, useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import "./style.css";
 import { RentItem } from 'src/types';
 import { useNavigate } from 'react-router';
-import { AUTH_ABSOLUTE_PATH, MYPAGE_DETAILS_ABSOLUTE_PATH, RENT_DETAIL_COUNT_PER_PAGE, RENT_DETAIL_COUNT_PER_SECTION } from 'src/constants';
+import { AUTH_ABSOLUTE_PATH, COUNT_PER_SECTION, MYPAGE_DETAILS_ABSOLUTE_PATH, USER_RENT_LIST_COUNT_PER_PAGE } from 'src/constants';
 import ResponseDto from 'src/apis/response.dto';
 import { GetMyRentPageResponseDto } from 'src/apis/payment/dto/response';
 import { getMyRentPageRequest, patchRentStatusRequest } from 'src/apis/payment';
 import { useCookies } from 'react-cookie';
 import { useUserStore } from 'src/stores';
+import { usePagination } from 'src/hooks';
 
 //                    component                    //
 function RentListItem({
@@ -93,59 +94,10 @@ export default function UserRentList() {
     const [cookies] = useCookies();
     const { loginUserRole } = useUserStore();
 
-    const [rentList, setRentList] = useState<RentItem[]>([]);
-    const [viewList, setViewList] = useState<RentItem[]>([]);
-    const [totalLength, setTotalLength] = useState<number>(0);
-
-    const [currentPage, setCurrentPage] = useState<number>(1);
-    const [currentSection, setCurrentSection] = useState<number>(1);
-    const [pageList, setPageList] = useState<number[]>([1]);
-    const [totalSection, setTotalSection] = useState<number>(1);
-    const [totalPage, setTotalPage] = useState<number>(1);
+    const {currentPage, viewList, pageList, setCurrentPage, setCurrentSection, changeBoardList,  onPageClickHandler, onPreSectionClickHandler, onPrePageClickHandler, onNextPageClickHandler, onNextSectionClickHandler} = usePagination<RentItem>(USER_RENT_LIST_COUNT_PER_PAGE, COUNT_PER_SECTION);
 
     //                    function                    //
     const navigator = useNavigate();
-
-    const changePage = (rentList: RentItem[], totalLength: number) => {
-        if (!currentPage) return;
-        const startIndex = (currentPage - 1) * RENT_DETAIL_COUNT_PER_PAGE;
-        let endIndex = currentPage * RENT_DETAIL_COUNT_PER_PAGE;
-        if (endIndex > totalLength - 1) endIndex = totalLength;
-        const viewList = rentList.slice(startIndex, endIndex);
-        setViewList(viewList);
-    };
-
-    const changeSection = (totalPage: number) => {
-        if (!currentSection) return;
-        const startPage = (currentSection * RENT_DETAIL_COUNT_PER_SECTION) - (RENT_DETAIL_COUNT_PER_SECTION - 1);
-        let endPage = currentSection * RENT_DETAIL_COUNT_PER_SECTION;
-        if (endPage > totalPage) endPage = totalPage;
-        const pageList: number[] = [];
-        for (let page = startPage; page <= endPage; page++) pageList.push(page);
-        setPageList(pageList);
-    };
-
-    const changeRentList = (rentList: RentItem[]) => {
-        if (!rentList || rentList.length === 0) {
-            alert('대여 목록을 찾을 수 없습니다.');
-            return;
-        }
-
-        setRentList(rentList);
-
-        const totalLength = rentList.length;
-        setTotalLength(totalLength);
-
-        const totalPage = Math.floor((totalLength - 1) / RENT_DETAIL_COUNT_PER_PAGE) + 1;
-        setTotalPage(totalPage);
-
-        const totalSection = Math.floor((totalPage - 1) / RENT_DETAIL_COUNT_PER_SECTION) + 1;
-        setTotalSection(totalSection);
-
-        changePage(rentList, totalLength);
-
-        changeSection(totalPage);
-    };
 
     const getMyPageRentListResponse = (result : GetMyRentPageResponseDto | ResponseDto | null) =>{ 
         if (!result) {
@@ -169,44 +121,11 @@ export default function UserRentList() {
             alert('대여 목록을 찾을 수 없습니다.');
             return;
         }    
-        changeRentList(rentList);
+        changeBoardList(rentList);
 
         setCurrentPage(!rentList.length ? 0 : 1);
         setCurrentSection(!rentList.length ? 0 : 1);
     }
-
-    //                    event handler                    //
-    const onPreSectionClickHandler = () => {
-        if (currentSection <= 1) return;
-        setCurrentSection(currentSection - 1);
-        setCurrentPage((currentSection - 1) * RENT_DETAIL_COUNT_PER_SECTION);
-    };
-
-    const onNextSectionClickHandler = () => {
-        if (currentSection === totalSection) return;
-        setCurrentSection(currentSection + 1);
-        setCurrentPage(currentSection * RENT_DETAIL_COUNT_PER_SECTION + 1);
-    };
-
-    const onPrePageClickHandler = () => {
-        if (currentPage <= 1) return;
-        if ((currentPage - 1) % RENT_DETAIL_COUNT_PER_SECTION === 0) {
-            setCurrentSection(currentSection - 1);
-        }
-        setCurrentPage(currentPage - 1);
-    };
-
-    const onNextPageClickHandler = () => {
-        if (currentPage === totalPage) return;
-        if (currentPage % RENT_DETAIL_COUNT_PER_SECTION === 0) {
-            setCurrentSection(currentSection + 1);
-        }
-        setCurrentPage(currentPage + 1);
-    };
-
-    const onPageClickHandler = (page: number) => {
-        setCurrentPage(page);
-    };
 
     //                    effect                    //
     useEffect(() => {
@@ -217,16 +136,6 @@ export default function UserRentList() {
         getMyRentPageRequest(cookies.accessToken).then(getMyPageRentListResponse);
 
     }, [cookies.accessToken, loginUserRole, navigator]);
-
-    useEffect(() => {
-        if (!rentList.length) return;
-        changePage(rentList, totalLength);
-    }, [currentPage, rentList, totalLength]);
-    
-    useEffect(() => {
-        if (!rentList.length) return;
-        changeSection(totalPage);
-    }, [currentSection, rentList.length, totalPage]);
 
     //                    render                    //
     return (
